backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
python-dotenv
google-generativeai
requests
joblib
matplotlib
seaborn